right registration, login using component
