+++ twitter-front/src/app/login-sign-up/login-sign-up.component.html
@@ -1,23 +1,32 @@
-<div class="button-container">
-  <button class = "switch-button"type="button" (click)="toggleForm()">{{ isLogin ? 'Sign Up' : 'Login' }}</button>
-</div>
-<form class="login-form" *ngIf="isLogin" (ngSubmit)="onLoginSubmit(username.value, password.value)">
-  <label>Username:</label>
-  <input type="text" #username required placeholder="z_svanov@kbtu.kz">
-
-  <label>Password:</label>
-  <input type="password" #password required placeholder="password">
-  <button type="submit">Submit</button>
-</form>
-
-<form class="signup-form" *ngIf="!isLogin" (ngSubmit)="onSignupSubmit(username.value, password.value, email.value)">
-  <label>Username:</label>
-  <input type="text" #username required placeholder="chiefzhantr">
-
-  <label>Password:</label>
-  <input type="password" #password required placeholder="password">
-  <label>Email:</label>
-  <input type="email" #email required placeholder="z_svanov@kbtu.kz">
-  <button type="submit">Submit</button>
-</form>
-
+<ng-container *ngIf="!logged">
+  <div class="button-is-login">
+    <button class = "button-is-login"type="button" (click)="toggleForm()">{{ isLogin ? 'Sign Up' : 'Login' }}</button>
+  </div>
+  <div class="login-form" *ngIf="isLogin">
+    <label>Username:</label>
+    <input type="text" [(ngModel)]="username">
+    <label>Password:</label>
+    <input type="password" [(ngModel)]="password">
+    <button (click)="login()">Login</button>
+  </div>
+  <div class="signup-form" *ngIf="!isLogin">
+    <label>First Name:</label>
+    <input type="text" [(ngModel)]="firstname">
+    <label>Last Name:</label>
+    <input type="text" [(ngModel)]="lastname">
+    <label>User:</label>
+    <input type="text" [(ngModel)]="username">
+    <label>Password:</label>
+    <input type="text" [(ngModel)]="password">
+    <label>Email:</label>
+    <input type="email" [(ngModel)]="email">
+    <label>Phone number:</label>
+    <input type="text" [(ngModel)]="phone">
+    <button type="submit" (click)="pushedRegister()">Register</button>
+    <form  *ngIf="want_to_register" (ngSubmit)="verify()">
+      <label>Verification code</label>
+      <input type="text" name="code" [(ngModel)]="codeTyped">
+      <button type="submit">Confirm</button>
+    </form>
+  </div>
+</ng-container>
