+++ twitter-front/src/app/news/news.component.html
@@ -1,4 +1,5 @@
 <app-searchbar></app-searchbar>
+<app-sidebar></app-sidebar>
 <body>
 <h3>What’s happening</h3>
 <ul>
