+++ twitter-front/src/app/profile/profile.component.html
@@ -1,5 +1,6 @@
 
 <app-searchbar></app-searchbar>
+<app-sidebar></app-sidebar>
 <body>
 <div class="flex-container">
   <img class="image-profile" src="../../assets/free-icon-profile-6917077.png" width="250" height="250">
