+++ twitter-front/src/app/settings/settings.component.html
@@ -1,4 +1,6 @@
+<app-sidebar></app-sidebar>
 <head>
+
   <title>Settings</title>
 </head>
 <body>
