+++ twitter-front/src/app/sidebar/sidebar.component.html
@@ -4,7 +4,7 @@
     <li class="element-li"><a (click)="showNews()">News</a>&nbsp;&nbsp;<img class = "image-item" src="https://cdn-icons-png.flaticon.com/512/81/81460.png" width="30px" height="30px" alt="news"></li>
     <li class="element-li"><a (click)="showProfile()">Profile</a>&nbsp;&nbsp;<img class = "image-item" src="https://cdn-icons-png.flaticon.com/512/10104/10104044.png" width="30px" height="30px" alt="news"></li>
     <li class="element-li"><a (click)="showSettings()">Settings</a>&nbsp;&nbsp;<img class = "image-item" src="https://cdn-icons-png.flaticon.com/512/3986/3986970.png" width="30px" height="30px" alt="news"></li>
-
+    <li class="element-li"><a (click)="logout()">Log out</a>&nbsp;&nbsp;<img class = "image-item" src="https://cdn-icons-png.flaticon.com/512/992/992680.png" width="30px" height="30px" alt="news"></li>
   </ul>
 </div>
 </body>
