+++ twitter-front/src/app/sidebar/sidebar.component.ts
@@ -22,4 +22,10 @@ export class SidebarComponent {
   showSettings() {
     this.router.navigate(['settings'])
   }
+  logout(){
+    localStorage.removeItem('token');
+    localStorage.removeItem('id');
+    localStorage.removeItem('username');
+    this.router.navigate(['login-signup'])
+  }
 }
